scoring_dicow/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scoring-dicow"
version = "0.1.0"
description = "Standalone scoring toolkit for DiCoW outputs against testset_tse datasets."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "meeteval>=0.4.3",
  "more-itertools>=10.0.0",
  "pyannote.core==5.0.0",
  "pyannote.metrics==3.2.1",
  "PyYAML>=6.0",
  "regex>=2024.0.0",
]

[project.scripts]
score-dicow = "scoring_dicow.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
scoring_dicow = ["text_norm/*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
